Finished implementation of take_action
Added the logic to check if the board has become separated into
multiple regions so that marbles are captured if they become
isolated on a separated region. Also made it so that a placement
action does not have to remove a ring from the board if no rings
may be removed.